+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Light Volume component placement now precisely estimated (calculated with polygon geometries).
+
 ## 1.7.6-beta.4 - 2026-08-01
 ### Fixed
 - Luminance don't have effects on VRCLightVolume shaders.